Добавляет финальную модуляцию в проект
