+++ js/main.js
@@ -1,6 +1,6 @@
-import {descriptionPhotos} from './data.js';
+import { getArrayPhotoDescriptions } from './createArrayPhotoDescriptions.js';
 
 // eslint-disable-next-line
 console.log(
-  descriptionPhotos()
+  getArrayPhotoDescriptions()
 );
