+++ js/util.js
@@ -18,4 +18,4 @@ const createIdGenerator = () => {
   };
 };
 
-export {getRandomInteger, getRandomArrayElement, createIdGenerator};
+export { getRandomInteger, getRandomArrayElement, createIdGenerator };
